looping over each token position for reward. Bfloat16/

+++ src/arena_capstone/algorithm/embedding_model.py
@@ -77,7 +77,7 @@ class EmbeddingFriendlyForCausalLM(EmbeddingFriendlyModel):
             we = tokens_or_onehot @ wte.weight
         else:
             we = wte(tokens_or_onehot)
-        return we.unsqueeze(0).half()
+        return we.unsqueeze(0).bfloat16()
 
     def forward_from_embed(self, embed):
         return self.model(inputs_embeds=embed)
